refactor: move Ethereum-related modules to `mpc-chain-ethereum`
