+++ Cargo.toml
@@ -8,7 +8,8 @@ members = [
     "chain-signatures/primitives",
     "integration-tests", 
     "chain-signatures/chain-integration-core", 
-    "chain-signatures/chain-canton"
+    "chain-signatures/chain-canton", 
+    "chain-signatures/chain-ethereum"
 ]
 exclude = ["infra/scripts/generate_keys"]
 resolver = "2"
@@ -55,6 +56,7 @@ tracing-subscriber = { version = "0.3.20", default-features = false, features = 
     "fmt",
 ] }
 url = { version = "2.4.0", features = ["serde"] }
+wasmtimer = "0.4.3"
 anchor-client = { version = "0.31.0", features = ["async"] }
 anchor-lang = "0.31.0"
 solana-sdk = "2.3.1"
@@ -81,6 +83,7 @@ mpc-contract = { path = "./chain-signatures/contract" }
 mpc-crypto = { path = "./chain-signatures/crypto" }
 mpc-chain-integration-core = { path = "./chain-signatures/chain-integration-core" }
 mpc-chain-canton = { path = "./chain-signatures/chain-canton" }
+mpc-chain-ethereum = { path = "./chain-signatures/chain-ethereum" }
 mpc-keys = { path = "./chain-signatures/keys" }
 mpc-node = { path = "./chain-signatures/node" }
 mpc-primitives = { path = "./chain-signatures/primitives" }
