+++ chain-signatures/chain-ethereum/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "mpc-chain-ethereum"
+edition = "2021"
+version.workspace = true
+
+[dependencies]
+alloy.workspace = true
+alloy-sol-types.workspace = true
+alloy-signer-local = "1.0.38"
+# TODO: replace with thiserror
+anyhow.workspace = true
+async-trait.workspace = true
+futures-util.workspace = true
+helios = { git = "https://github.com/a16z/helios", rev = "e1f9a50f73d7af6f6ae4acf2b663d04aab6adb19", optional = true }
+hex.workspace = true
+k256.workspace = true
+mpc-chain-integration-core.workspace = true
+mpc-crypto.workspace = true
+mpc-primitives.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+reqwest.workspace = true
+tracing.workspace = true
+
+[features]
+helios = ["dep:helios"]
+
+[dev-dependencies]
+mockito.workspace = true
