+++ chain-signatures/chain-ethereum/src/util.rs
@@ -1,0 +1,8 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+
+pub fn current_unix_timestamp() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("Time went backwards")
+        .as_secs()
+}
